requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pyyaml>=6.0
python-frontmatter>=1.1.0
httpx>=0.27.0
anthropic>=0.49.0
